added .export_csv() method under Results class
just exporting the results as csv that we can retrieve from the query function. I believe this will be useful for developers. sure, a user could do this on their own, but it's much easier to just call results.export_csv(). also, imported time, so we can have a little time stamp at the end of the file, to avoid overwriting conflicts.